content-security-policy/plugin-types/plugintypes-nourl-blocked.html

<!DOCTYPE html>
<html>

<head>
  <meta http-equiv="Content-Security-Policy" content="plugin-types text/plain;">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
  <script>
    var t = async_test('Should not load the object because it does not match plugin-types');
    window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
      assert_equals(e.violatedDirective, "plugin-types");
      assert_equals(e.blockedURI, "");
    }));
  </script>

  <object type="application/x-webkit-test-netscape"></object>
</body>

</html>
